run project on current go
