+++ examples/mic.go
@@ -2,28 +2,27 @@
  * Uses the microphone to record chunks and runs the YIN Pitch Detection on the chunks.
  * Writes the output to a channel of floats
  * Ctrl+C to break the audio recording loop
- * 
+ *
  * Requires portaudio for go : go get github.com/gordonklaus/portaudio
 
  */
- 
-
-
-
 
 package main
 
-import "github.com/mrnikho/yingo"
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/mrnikho/yingo"
+)
 
 func main() {
-	
+
 	pitchChannel := make(chan float32)
-	yingo.MicInput(2048, true, pitchChannel)
-	
+	yingo.MicInput(2048, true, &pitchChannel)
+
 	for {
-	select {
-		case pitch:= <- pitchChannel:
+		select {
+		case pitch := <-pitchChannel:
 			fmt.Println("Pitch is: ", pitch)
 		}
 	}
