+++ examples/mono-wav/mono-wav.go
