+++ examples/test/test.go
@@ -1,0 +1,136 @@
+/*Ported from here: https://github.com/ashokfernandez/Yin-Pitch-Tracking
+ *
+ */
+
+package main
+
+import (
+	"fmt"
+
+	"github.com/mrnikho/yingo"
+)
+
+func main() {
+
+	data := []int16{
+		19, 24, 29, 34, 38, 42, 46, 49, 52, 54, 56, 58, 59, 60, 61, 61, 60,
+		60, 59, 58, 56, 54, 52, 49, 47, 44, 41, 39, 36, 34, 32, 29, 27, 24,
+		21, 18, 15, 12, 9, 5, 2, 0, -3, -6, -9, -12, -14, -17, -19, -21,
+		-23, -25, -27, -29, -31, -33, -35, -37, -39, -41, -43, -44, -45,
+		-46, -47, -48, -48, -48, -48, -47, -46, -45, -44, -43, -41, -40,
+		-38, -37, -35, -34, -32, -31, -30, -28, -27, -26, -24, -23, -22,
+		-21, -20, -18, -18, -17, -17, -17, -17, -18, -19, -21, -22, -24,
+		-26, -28, -30, -31, -33, -35, -37, -40, -42, -44, -47, -50, -52,
+		-56, -59, -63, -67, -71, -75, -79, -82, -86, -89, -92, -95, -98,
+		-101, -103, -105, -107, -109, -111, -112, -114, -115, -117, -118,
+		-120, -121, -122, -122, -123, -123, -123, -123, -122, -121, -120,
+		-119, -117, -116, -114, -112, -109, -107, -104, -101, -98, -95, -92,
+		-89, -85, -81, -77, -73, -69, -65, -61, -56, -52, -47, -43, -39,
+		-35, -31, -28, -24, -20, -16, -12, -8, -4, 0, 3, 7, 11, 15, 20, 24,
+		28, 32, 36, 40, 44, 47, 51, 54, 57, 60, 62, 64, 65, 67, 68, 70, 71,
+		73, 74, 76, 77, 79, 81, 83, 85, 88, 90, 92, 95, 97, 99, 101, 103,
+		104, 105, 106, 106, 106, 106, 105, 104, 104, 103, 102, 102, 101,
+		102, 102, 103, 104, 106, 107, 109, 111, 112, 114, 115, 115, 116,
+		116, 115, 114, 113, 112, 111, 109, 108, 107, 106, 105, 105, 105,
+		105, 106, 107, 108, 109, 110, 112, 112, 113, 113, 113, 113, 112,
+		111, 110, 109, 107, 105, 104, 103, 101, 100, 99, 99, 98, 98, 98, 98,
+		97, 97, 97, 96, 95, 94, 93, 91, 90, 89, 87, 85, 84, 82, 81, 80, 79,
+		78, 76, 75, 74, 73, 71, 70, 68, 66, 64, 63, 61, 59, 57, 56, 54, 52,
+		51, 49, 48, 47, 46, 46, 45, 45, 44, 44, 43, 42, 41, 40, 38, 37, 35,
+		33, 31, 29, 27, 25, 23, 22, 20, 19, 18, 17, 17, 17, 17, 17, 17, 18,
+		18, 19, 19, 20, 19, 19, 18, 17, 16, 14, 12, 11, 9, 8, 6, 5, 4, 3, 2,
+		2, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, -2, -3, -5, -6, -8, -9,
+		-10, -12, -13, -13, -14, -15, -15, -16, -16, -17, -17, -18, -19,
+		-20, -21, -22, -23, -24, -25, -26, -27, -27, -28, -29, -29, -30,
+		-31, -32, -32, -33, -34, -34, -35, -35, -36, -36, -37, -37, -38,
+		-38, -39, -39, -39, -39, -39, -39, -39, -40, -40, -40, -40, -40,
+		-39, -39, -39, -38, -38, -38, -37, -37, -37, -38, -38, -39, -40,
+		-41, -42, -43, -44, -44, -45, -46, -46, -47, -47, -48, -49, -50,
+		-51, -53, -55, -58, -61, -64, -67, -70, -73, -76, -79, -83, -86,
+		-90, -93, -97, -101, -105, -109, -112, -115, -117, -119, -121, -122,
+		-123, -123, -123, -123, -122, -121, -119, -118, -116, -115, -113,
+		-112, -109, -106, -103, -99, -94, -90, -85, -80, -75, -69, -64, -58,
+		-52, -46, -40, -33, -27, -20, -14, -8, -1, 4, 10, 16, 22, 27, 33,
+		39, 44, 49, 54, 59, 63, 67, 70, 72, 74, 76, 77, 78, 78, 78, 77, 76,
+		75, 73, 72, 70, 68, 66, 64, 62, 59, 57, 55, 53, 50, 47, 43, 40, 35,
+		31, 27, 22, 18, 14, 10, 6, 2, 0, -4, -7, -10, -14, -17, -21, -24,
+		-28, -32, -35, -38, -41, -43, -46, -48, -50, -51, -52, -52, -52,
+		-52, -51, -50, -50, -49, -47, -46, -45, -44, -44, -43, -42, -41,
+		-40, -39, -38, -36, -34, -33, -31, -29, -28, -26, -24, -23, -22,
+		-21, -21, -20, -20, -20, -20, -21, -22, -23, -25, -27, -29, -32,
+		-35, -38, -41, -44, -48, -51, -55, -58, -62, -66, -70, -73, -77,
+		-81, -85, -88, -92, -95, -99, -103, -106, -110, -113, -116, -119,
+		-122, -124, -127, -130, -133, -136, -139, -141, -144, -146, -148,
+		-149, -150, -151, -151, -151, -151, -151, -150, -149, -148, -146,
+		-144, -142, -139, -137, -134, -131, -129, -126, -123, -120, -117,
+		-114, -111, -108, -104, -101, -98, -94, -91, -87, -82, -77, -72,
+		-67, -61, -55, -48, -42, -36, -30, -24, -18, -13, -8, -3, 0, 3, 6,
+		9, 12, 14, 17, 20, 23, 26, 30, 34, 38, 43, 48, 52, 57, 62, 67, 71,
+		74, 78, 81, 83, 85, 86, 87, 87, 87, 87, 88, 89, 90, 91, 92, 94, 96,
+		98, 100, 102, 105, 107, 109, 111, 113, 114, 115, 116, 116, 116, 116,
+		115, 115, 115, 114, 114, 114, 114, 114, 114, 115, 115, 116, 117,
+		117, 117, 118, 118, 118, 118, 118, 118, 118, 118, 118, 118, 118,
+		118, 118, 118, 118, 118, 118, 117, 117, 117, 117, 116, 116, 116,
+		115, 115, 114, 113, 112, 111, 110, 109, 108, 108, 107, 107, 106,
+		106, 106, 106, 106, 106, 105, 105, 104, 103, 101, 100, 98, 96, 94,
+		92, 90, 88, 86, 85, 83, 81, 80, 78, 77, 76, 75, 75, 75, 75, 75, 74,
+		74, 73, 72, 71, 70, 68, 67, 65, 63, 60, 58, 56, 54, 52, 50, 48, 47,
+		46, 46, 45, 45, 44, 44, 44, 43, 43, 43, 42, 41, 41, 40, 39, 38, 38,
+		37, 36, 36, 35, 34, 34, 33, 32, 32, 31, 30, 29, 29, 28, 28, 28, 27,
+		27, 27, 26, 25, 25, 24, 23, 22, 21, 20, 20, 19, 19, 18, 17, 17, 16,
+		16, 15, 15, 14, 13, 12, 11, 9, 8, 7, 5, 4, 3, 3, 2, 1, 0, 0, 0, -1,
+		-1, -2, -3, -3, -4, -6, -7, -8, -10, -11, -12, -13, -14, -16, -17,
+		-18, -19, -20, -21, -22, -23, -24, -25, -25, -25, -25, -26, -26,
+		-26, -26, -27, -27, -28, -29, -29, -30, -31, -31, -32, -33, -34,
+		-34, -35, -35, -35, -35, -35, -35, -34, -34, -34, -33, -33, -33,
+		-33, -33, -33, -34, -34, -35, -36, -38, -40, -42, -45, -48, -51,
+		-53, -56, -59, -62, -65, -69, -72, -75, -79, -82, -85, -88, -91,
+		-95, -99, -104, -108, -111, -115, -118, -122, -125, -127, -130,
+		-131, -133, -134, -134, -135, -135, -135, -134, -134, -133, -131,
+		-130, -129, -127, -125, -123, -121, -117, -113, -108, -103, -98,
+		-92, -86, -79, -73, -67, -62, -56, -51, -45, -39, -33, -26, -19,
+		-13, -6, 0, 5, 11, 17, 22, 27, 31, 35, 39, 42, 44, 46, 48, 49, 51,
+		52, 53, 54, 54, 55, 55, 55, 54, 54, 53, 52, 51, 50, 48, 46, 44, 41,
+		38, 34, 30, 26, 22, 17, 13, 8, 4, 0, -5, -9, -14, -18, -22, -25,
+		-29, -32, -36, -39, -42, -45, -48, -51, -54, -56, -58, -60, -61,
+		-63, -64, -65, -66, -66, -67, -67, -67, -67, -66, -66, -66, -65,
+		-64, -63, -62, -60, -58, -56, -53, -50, -48, -45, -42, -40, -38,
+		-36, -33, -32, -30, -28, -27, -26, -26, -26, -25, -25, -26, -27,
+		-28, -30, -32, -35, -38, -41, -44, -47, -50, -52, -55, -58, -61,
+		-63, -66, -68, -71, -74, -77, -81, -86, -90, -95, -99, -103, -107,
+		-111, -115, -119, -123, -127, -130, -133, -136, -138, -141, -143,
+		-144, -145, -145, -144, -144, -143, -143, -143, -143, -143, -143,
+		-143, -143, -144, -144, -145, -145, -144, -143, -141, -138, -135,
+		-131, -127, -123, -118, -113, -108, -103, -98, -93, -89, -85, -81,
+		-77, -73, -69, -65, -61, -57, -53, -49, -45, -40, -35, -30, -25,
+		-20, -15, -10, -4, 0, 5, 11, 16, 21, 26, 30, 34, 38, 42, 45, 49, 52,
+		55, 58, 61, 64, 67, 70, 73, 76, 79, 82, 84, 87, 89, 92, 94, 96, 98,
+		99, 101, 103, 104, 106, 108, 109, 111, 112, 113, 115, 116, 118, 119,
+		120, 121, 121, 122, 122, 123, 123, 123, 123, 123, 123, 123, 123,
+		123, 123, 123, 123, 124, 124, 125, 125, 126, 127, 128, 128, 129,
+		129, 129, 129, 128, 127, 127, 126, 124, 123, 122, 121, 120, 119,
+		119, 119, 119, 119, 119, 119, 119, 120, 120, 120, 120, 120, 119,
+		119, 118, 117, 116, 114, 113, 112, 110, 109, 107, 105, 104, 102,
+		101, 99, 98, 97, 96, 95, 94, 93, 93, 92, 91, 90, 89, 88, 87, 86, 84,
+		83, 81, 79, 78, 76, 74, 73, 71, 70, 69, 67, 66, 64, 62, 61, 59, 58,
+		56, 55, 53, 51, 50, 49, 48, 47, 46, 45, 44, 43, 42, 42, 41, 40, 39,
+		38, 37, 36, 35, 34, 33, 32, 32, 31, 30, 29, 28, 27, 26, 25, 24, 23,
+		23, 22, 22, 22, 21, 21, 20, 20, 19, 19, 18, 17, 17, 16, 15, 13, 12, 11, 9, 8, 7, 6,
+		5, 5, 4, 3, 2, 1, 1, 0, 0, 0, 0, -1, -2, -3, -4, -4, -5, -7, -8, -10, -11, -13, -14, -16, -17,
+		-19, -20, -22, -23, -23, -23, -24, -24, -24, -25, -25, -26, -27, -28}
+
+	buffersize := 100
+	threshold := float32(0.05)
+
+	yin := yingo.Yin{}
+
+	var pitch float32
+
+	for pitch < 10 {
+		yin.YinInit(buffersize, threshold)
+		pitch = yin.GetPitch(&data)
+		buffersize += 1
+	}
+
+	fmt.Printf("Estimated pitch is %f Hz with buffer length %d and probability %f\n", pitch, yin.BufferSize, yin.GetProb())
+}
