+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/mrnikho/yingo
+
+go 1.19
+
+require github.com/gordonklaus/portaudio v0.0.0-20221027163845-7c3b689db3cc
